requirements.txt
python-dotenv
openai
llama-index
pypdf
streamlit
nltk

